feat(analyze): separte logic
